+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2023-01-5
+
+### Fixed
+
+- Image export wasn't working. This is now fixed.
+
 ## 2023-01-04
 
 ### Added